+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Use new service config handling for database configuration
 ### Fixed
+- DF-1033 Correct datetime config option usage
 
 ## [0.8.2] - 2017-03-29
 ### Fixed